refactor: extract xtdb staging
